Struct graph_runtime_wasm::host_exports::HostExports

pub struct HostExports<C: Blockchain> {
    pub(crate) subgraph_id: DeploymentHash,
    pub(crate) api_version: Version,
    data_source_name: String,
    data_source_address: Vec<u8>,
    data_source_network: String,
    data_source_context: Arc<Option<DataSourceContext>>,
    causality_region: String,
    templates: Arc<Vec<C::DataSourceTemplate>>,
    pub(crate) link_resolver: Arc<dyn LinkResolver>,
    store: Arc<dyn SubgraphStore>,
}

Fields
subgraph_id: DeploymentHashapi_version: Versiondata_source_name: Stringdata_source_address: Vec<u8>data_source_network: Stringdata_source_context: Arc<Option<DataSourceContext>>causality_region: String

Some data sources have indeterminism or different notions of time. These need to be each be stored separately to separate causality between them, and merge the results later. Right now, this is just the ethereum networks but will be expanded for ipfs and the availability chain.

templates: Arc<Vec<C::DataSourceTemplate>>link_resolver: Arc<dyn LinkResolver>store: Arc<dyn SubgraphStore>
Implementations

impl<C: Blockchain> HostExports<C>

pub fn new(
subgraph_id: DeploymentHash,
data_source: &impl DataSource<C>,
data_source_network: String,
templates: Arc<Vec<C::DataSourceTemplate>>,
link_resolver: Arc<dyn LinkResolver>,
store: Arc<dyn SubgraphStore>
) -> Self

pub(crate) fn abort(
&self,
message: Option<String>,
file_name: Option<String>,
line_number: Option<u32>,
column_number: Option<u32>
) -> Result<Never, DeterministicHostError>

pub(crate) fn store_set(
&self,
logger: &Logger,
state: &mut BlockState<C>,
proof_of_indexing: &SharedProofOfIndexing,
entity_type: String,
entity_id: String,
data: HashMap<String, Value>,
stopwatch: &StopwatchMetrics
) -> Result<(), Error>

pub(crate) fn store_remove(
&self,
logger: &Logger,
state: &mut BlockState<C>,
proof_of_indexing: &SharedProofOfIndexing,
entity_type: String,
entity_id: String
) -> Result<(), HostExportError>

pub(crate) fn store_get(
&self,
state: &mut BlockState<C>,
entity_type: String,
entity_id: String
) -> Result<Option<Entity>, Error>

pub(crate) fn big_int_to_hex(
&self,
n: BigInt
) -> Result<String, DeterministicHostError>

Prints the module of n in hex. Integers are encoded using the least amount of digits (no leading zero digits). Their encoding may be of uneven length. The number zero encodes as “0x0”.

https://godoc.org/github.com/ethereum/go-ethereum/common/hexutil#hdr-Encoding_Rules

pub(crate) fn ipfs_cat(
&self,
logger: &Logger,
link: String
) -> Result<Vec<u8>, Error>

pub(crate) fn ipfs_map(
link_resolver: &Arc<dyn LinkResolver>,
module: &mut WasmInstanceContext<C>,
link: String,
callback: &str,
user_data: Value,
flags: Vec<String>
) -> Result<Vec<BlockState<C>>, Error>

pub(crate) fn json_to_i64(
&self,
json: String
) -> Result<i64, DeterministicHostError>

Expects a decimal string.

pub(crate) fn json_to_u64(
&self,
json: String
) -> Result<u64, DeterministicHostError>

Expects a decimal string.

pub(crate) fn json_to_f64(
&self,
json: String
) -> Result<f64, DeterministicHostError>

Expects a decimal string.

pub(crate) fn json_to_big_int(
&self,
json: String
) -> Result<Vec<u8>, DeterministicHostError>

Expects a decimal string.

pub(crate) fn crypto_keccak_256(
&self,
input: Vec<u8>
) -> Result<[u8; 32], DeterministicHostError>

pub(crate) fn big_int_plus(
&self,
x: BigInt,
y: BigInt
) -> Result<BigInt, DeterministicHostError>

pub(crate) fn big_int_minus(
&self,
x: BigInt,
y: BigInt
) -> Result<BigInt, DeterministicHostError>

pub(crate) fn big_int_times(
&self,
x: BigInt,
y: BigInt
) -> Result<BigInt, DeterministicHostError>

pub(crate) fn big_int_divided_by(
&self,
x: BigInt,
y: BigInt
) -> Result<BigInt, DeterministicHostError>

pub(crate) fn big_int_mod(
&self,
x: BigInt,
y: BigInt
) -> Result<BigInt, DeterministicHostError>

pub(crate) fn big_int_pow(
&self,
x: BigInt,
exponent: u8
) -> Result<BigInt, DeterministicHostError>

Limited to a small exponent to avoid creating huge BigInts.

pub(crate) fn big_int_from_string(
&self,
s: String
) -> Result<BigInt, DeterministicHostError>

pub(crate) fn big_int_bit_or(
&self,
x: BigInt,
y: BigInt
) -> Result<BigInt, DeterministicHostError>

pub(crate) fn big_int_bit_and(
&self,
x: BigInt,
y: BigInt
) -> Result<BigInt, DeterministicHostError>

pub(crate) fn big_int_left_shift(
&self,
x: BigInt,
bits: u8
) -> Result<BigInt, DeterministicHostError>

pub(crate) fn big_int_right_shift(
&self,
x: BigInt,
bits: u8
) -> Result<BigInt, DeterministicHostError>

pub(crate) fn bytes_to_base58(
&self,
bytes: Vec<u8>
) -> Result<String, DeterministicHostError>

Useful for IPFS hashes stored as bytes

pub(crate) fn big_decimal_plus(
&self,
x: BigDecimal,
y: BigDecimal
) -> Result<BigDecimal, DeterministicHostError>

pub(crate) fn big_decimal_minus(
&self,
x: BigDecimal,
y: BigDecimal
) -> Result<BigDecimal, DeterministicHostError>

pub(crate) fn big_decimal_times(
&self,
x: BigDecimal,
y: BigDecimal
) -> Result<BigDecimal, DeterministicHostError>

pub(crate) fn big_decimal_divided_by(
&self,
x: BigDecimal,
y: BigDecimal
) -> Result<BigDecimal, DeterministicHostError>

Maximum precision of 100 decimal digits.

pub(crate) fn big_decimal_equals(
&self,
x: BigDecimal,
y: BigDecimal
) -> Result<bool, DeterministicHostError>

pub(crate) fn big_decimal_to_string(
&self,
x: BigDecimal
) -> Result<String, DeterministicHostError>

pub(crate) fn big_decimal_from_string(
&self,
s: String
) -> Result<BigDecimal, DeterministicHostError>

pub(crate) fn data_source_create(
&self,
logger: &Logger,
state: &mut BlockState<C>,
name: String,
params: Vec<String>,
context: Option<DataSourceContext>,
creation_block: BlockNumber
) -> Result<(), HostExportError>

pub(crate) fn ens_name_by_hash(
&self,
hash: &str
) -> Result<Option<String>, Error>

pub(crate) fn log_log(
&self,
logger: &Logger,
level: Level,
msg: String
) -> Result<(), DeterministicHostError>

pub(crate) fn data_source_address(&self) -> Vec<u8>

pub(crate) fn data_source_network(&self) -> String

pub(crate) fn data_source_context(&self) -> Entity