Module graph_runtime_wasm::mapping

Public interface of the crate, receives triggers to be processed. Pre-processes modules and manages their threads. Serves as an interface from host to module.

Structs

LOG_TRIGGER_DATA
Verbose logging of mapping inputs

MappingContext

MappingRequest

ValidModule
A pre-processed and valid WASM module, ready to be started as a WasmModule.

Functions

spawn_module
Spawn a wasm module in its own thread.