Struct graph_runtime_wasm::module::WasmInstance

pub struct WasmInstance<C: Blockchain> {
    instance: Instance,
    instance_ctx: Rc<RefCell<Option<WasmInstanceContext<C>>>>,
}

Handle to a WASM instance, which is terminated if and only if this is dropped.

Fields
instance: Instanceinstance_ctx: Rc<RefCell<Option<WasmInstanceContext<C>>>>
Implementations

impl<C: Blockchain> WasmInstance<C>

pub(crate) fn handle_json_callback(
self,
handler_name: &str,
value: &Value,
user_data: &Value
) -> Result<BlockState<C>, Error>

pub(crate) fn handle_trigger(
self,
trigger: C::MappingTrigger
) -> Result<BlockState<C>, MappingError>

pub fn take_ctx(&mut self) -> WasmInstanceContext<C>

pub(crate) fn instance_ctx(&self) -> Ref<'_, WasmInstanceContext<C>>

pub fn instance_ctx_mut(&self) -> RefMut<'_, WasmInstanceContext<C>>

pub fn get_func(&self, func_name: &str) -> Func

fn invoke_handler<T>(
&mut self,
handler: &str,
arg: AscPtr<T>
) -> Result<BlockState<C>, MappingError>

impl<C: Blockchain> WasmInstance<C>

pub fn from_valid_module_with_ctx(
valid_module: Arc<ValidModule>,
ctx: MappingContext<C>,
host_metrics: Arc<HostMetrics>,
timeout: Option<Duration>,
experimental_features: ExperimentalFeatures
) -> Result<WasmInstance<C>, Error>

Instantiates the module and sets it to be interrupted after timeout.